rewrites of script generator
