+++ Shared/LobbyClient/Battle.cs
@@ -127,216 +127,7 @@ namespace LobbyClient
         }
 
 
-        /// <summary>
-        /// Generates script
-        /// </summary>
-        /// <param name="playersExport">list of players</param>
-        /// <param name="localUser">myself</param>
-        /// <param name="loopbackListenPort">listen port for autohost interface</param>
-        /// <param name="zkSearchTag">hackish search tag</param>
-        /// <param name="startSetup">structure with custom extra data</param>
-        /// <returns></returns>
-        public string GenerateScript(out List<UserBattleStatus> playersExport,
-                                     User localUser,
-                                     int loopbackListenPort,
-                                     string zkSearchTag,
-                                     SpringBattleStartSetup startSetup)
-        {
-            var previousCulture = Thread.CurrentThread.CurrentCulture;
-            try
-            {
-                Thread.CurrentThread.CurrentCulture = CultureInfo.InvariantCulture;
-
-                playersExport = new List<UserBattleStatus>();
-                var isHost = localUser.Name == Founder.Name;
-
-                var myUbs = Users[localUser.Name];
-                if (!isHost)
-                {
-                    var sb = new StringBuilder();
-                    sb.AppendLine("[GAME]");
-                    sb.AppendLine("{");
-                    sb.AppendFormat("HostIP={0};\n", Ip);
-                    sb.AppendFormat("HostPort={0};\n", HostPort);
-                    sb.AppendLine("IsHost=0;");
-                    sb.AppendFormat("MyPlayerName={0};\n", localUser.Name);
-                    if (myUbs != null)
-                    {
-                        if (myUbs.ScriptPassword != null) sb.AppendFormat("MyPasswd={0};\n", myUbs.ScriptPassword);
-                    }
-                    else
-                    {
-                        sb.AppendFormat("MyPasswd={0};\n", localUser.Name); // used for mid-game join .. if no userbattlestatus, use own name
-                    }
-                    sb.AppendLine("}");
-                    return sb.ToString();
-                }
-                else
-                {
-
-                    var script = new StringBuilder();
-
-                    script.AppendLine("[GAME]");
-                    script.AppendLine("{");
-
-                    script.AppendFormat("   ZkSearchTag={0};\n", zkSearchTag);
-                    script.AppendFormat("  Mapname={0};\n", MapName);
-
-                    script.AppendFormat("  StartPosType=2;\n");
-
-                    script.AppendFormat("  GameType={0};\n", ModName);
-                    script.AppendFormat("  ModHash=1;\n");
-                    script.AppendFormat("  MapHash=1;\n");
-
-                    script.AppendFormat("  AutohostPort={0};\n", loopbackListenPort);
-                    script.AppendLine();
-                    script.AppendFormat("  HostIP={0};\n", Ip);
-                    script.AppendFormat("  HostPort={0};\n", HostPort);
-                    script.AppendFormat("  SourcePort={0};\n", 8300);
-                    script.AppendFormat("  IsHost=1;\n");
-                    script.AppendLine();
-
-                    //script.AppendFormat("  MyPlayerName={0};\n", localUser.Name);
-
-                    List<UserBattleStatus> users;
-                    List<BotBattleStatus> bots;
-
-                    if (startSetup != null && startSetup.BalanceTeamsResult != null && startSetup.BalanceTeamsResult.Players != null)
-                    {
-                        // if there is a balance results as a part of start setup, use values from this (override lobby state)
-                        users = Users.Values.ToList();
-                        bots = new List<BotBattleStatus>(this.Bots.Values.Select(x => (BotBattleStatus)x.Clone()));
-                        foreach (var p in startSetup.BalanceTeamsResult.Players)
-                        {
-                            var us = users.FirstOrDefault(x => x.Name == p.Name);
-                            if (us == null)
-                            {
-                                us = new UserBattleStatus(p.Name, new User() { AccountID = p.LobbyID }, Password); // TODO this "password" use does not look right
-                                users.Add(us);
-                            }
-                            us.TeamNumber = p.TeamID;
-                            us.IsSpectator = p.IsSpectator;
-                            us.AllyNumber = p.AllyID;
-
-                        }
-                        foreach (var p in startSetup.BalanceTeamsResult.Bots)
-                        {
-                            var bot = bots.FirstOrDefault(x => x.Name == p.BotName);
-                            if (bot == null)
-                            {
-                                bot = new BotBattleStatus(p.BotName, p.Owner, p.BotAI);
-                                bots.Add(bot);
-                            }
-                            bot.AllyNumber = bot.AllyNumber;
-                            bot.TeamNumber = bot.TeamNumber;
-                        }
-
-                        foreach (var u in users.Where(x => !startSetup.BalanceTeamsResult.Players.Any(y => y.Name == x.Name)))
-                        {
-                            u.IsSpectator = true;
-                        } // spec those not known at the time of balance
-                    }
-                    else
-                    {
-                        users = this.Users.Values.ToList();
-                        bots = this.Bots.Values.ToList();
-                    }
-
-
-                    GeneratePlayerSection(playersExport, users, script, bots, Rectangles, ModOptions, localUser, startSetup);
-
-                    return script.ToString();
-                }
-            }
-            finally
-            {
-                Thread.CurrentThread.CurrentCulture = previousCulture;
-            }
-        }
-
-        public static void GeneratePlayerSection(List<UserBattleStatus> playersExport,
-            List<UserBattleStatus> users,
-            StringBuilder script,
-            List<BotBattleStatus> bots,
-            IDictionary<int, BattleRect> _rectangles,
-            Dictionary<string, string> _modOptions,
-            User localUser = null,
-            SpringBattleStartSetup startSetup = null
-           )
-        {
-            // ordinary battle stuff
-
-            var userNum = 0;
-            var teamNum = 0;
-            var aiNum = 0;
-
-            //players is excluding self (so "springie doesn't appear as spec ingame") & excluding bots (bots is added later for each owner)
-            var non_botUsers = users.Where(u => !bots.Any(b => b.Name == u.Name)); //.OrderBy(x => x.TeamNumber);
-            if (localUser != null) //I am a server
-                non_botUsers = non_botUsers.Where(x => x.Name != localUser.Name);
-
-            foreach (var u in non_botUsers.OrderBy(x => x.TeamNumber))
-            {
-                ScriptAddUser(script, userNum, playersExport, startSetup, teamNum, u);
-
-                if (!u.IsSpectator)
-                {
-                    ScriptAddTeam(script, teamNum, userNum, u);
-                    teamNum++;
-                }
-
-                foreach (var b in bots.Where(x => x.owner == u.Name))
-                {
-                    ScriptAddBot(script, aiNum, teamNum, userNum, b);
-                    aiNum++;
-                    ScriptAddTeam(script, teamNum, userNum, b);
-                    teamNum++;
-                }
-                userNum++;
-            }
-
-            // ALLIANCES AND START BOXES
-            var startboxes = new StringBuilder();
-            startboxes.Append("return { ");
-            script.AppendLine();
-            for (int allyNumber = 0; allyNumber < Spring.MaxAllies; allyNumber++)
-            {
-                script.AppendFormat("[ALLYTEAM{0}]\n", allyNumber);
-                script.AppendLine("{");
-                script.AppendLine("     NumAllies=0;");
-                BattleRect rect;
-                if (_rectangles.TryGetValue(allyNumber, out rect)) {
-                    double left = 0, top = 0, right = 1, bottom = 1;
-                    rect.ToFractions(out left, out top, out right, out bottom);
-                    startboxes.AppendFormat(CultureInfo.InvariantCulture, "[{0}] = ", allyNumber);
-                    startboxes.Append("{ ");
-                    startboxes.AppendFormat(CultureInfo.InvariantCulture, "{0}, {1}, {2}, {3}", left, top, right, bottom);
-                    startboxes.Append(" }, ");
-                }
-                script.AppendLine("}");
-            }
-
-            startboxes.Append("}");
-            script.AppendLine();
-
-                script.AppendLine("  [MODOPTIONS]");
-                script.AppendLine("  {");
-
-                script.AppendFormat("    startboxes={0};\n", startboxes.ToString());
-
-                var options = new Dictionary<string, string>(_modOptions);
-
-                // replace/add custom modoptions from startsetup (if they exist)
-                if (startSetup != null && startSetup.ModOptions != null) foreach (var entry in startSetup.ModOptions) options[entry.Key] = entry.Value;
-
-                // write final options to script
-                foreach (var kvp in options) script.AppendFormat("    {0}={1};\n", kvp.Key, kvp.Value);
-
-                script.AppendLine("  }");
-
-
-            script.AppendLine("}");
-        }
+     
 
 
         public int GetFirstEmptyRectangle()
@@ -361,66 +152,6 @@ namespace LobbyClient
             return String.Format("{0} {1} ({2}+{3}/{4})", ModName, MapName, NonSpectatorCount, SpectatorCount, MaxPlayers);
         }
 
-        static void ScriptAddBot(StringBuilder script, int aiNum, int teamNum, int userNum, BotBattleStatus status)
-        {
-            // AI
-            var split = status.aiLib.Split('|');
-            script.AppendFormat("  [AI{0}]\n", aiNum);
-            script.AppendLine("  {");
-            script.AppendFormat("    Name={0};\n", status.Name);
-            script.AppendFormat("    ShortName={0};\n", split[0]);
-            script.AppendFormat("    Version={0};\n", split.Length > 1 ? split[1] : ""); //having no value is better. Related file: ResolveSkirmishAIKey() at Spring/ExternalAI/IAILibraryManager.cpp 
-            script.AppendFormat("    Team={0};\n", teamNum);
-            script.AppendFormat("    Host={0};\n", userNum);
-            script.AppendLine("    IsFromDemo=0;");
-            script.AppendLine("    [Options]");
-            script.AppendLine("    {");
-            script.AppendLine("    }");
-            script.AppendLine("  }\n");
-        }
-
-        static void ScriptAddTeam(StringBuilder script, int teamNum, int userNum, UserBattleStatus status)
-        {
-            // BOT TEAM
-            script.AppendFormat("  [TEAM{0}]\n", teamNum);
-            script.AppendLine("  {");
-            script.AppendFormat("     TeamLeader={0};\n", userNum);
-            script.AppendFormat("     AllyTeam={0};\n", status.AllyNumber);
-            //var side = "mission";
-            //script.AppendFormat("     Side={0};\n", mod.Sides[0]); // is this use of "mod" needed at all?
-
-            script.AppendFormat("     Handicap={0};\n", 0);
-            script.AppendLine("  }");
-        }
-
-        static void ScriptAddUser(StringBuilder script, int userNum, List<UserBattleStatus> playersExport, SpringBattleStartSetup startSetup, int teamNum, UserBattleStatus status)
-        {
-            var export = status.Clone();
-            export.TeamNumber = teamNum;
-            playersExport.Add(status);
-
-            // PLAYERS
-            script.AppendFormat("  [PLAYER{0}]\n", userNum);
-            script.AppendLine("  {");
-            script.AppendFormat("     Name={0};\n", status.Name);
-
-            script.AppendFormat("     Spectator={0};\n", status.IsSpectator ? 1 : 0);
-            if (!status.IsSpectator) script.AppendFormat("     Team={0};\n", teamNum);
-
-            if (status.LobbyUser != null)
-            {
-                script.AppendFormat("     CountryCode={0};\n", status.LobbyUser.Country);
-                script.AppendFormat("     LobbyID={0};\n", status.LobbyUser.AccountID);
-            }
-            if (status.ScriptPassword != null) script.AppendFormat("     Password={0};\n", status.ScriptPassword);
-
-            if (startSetup != null)
-            {
-                var entry = startSetup.UserParameters.FirstOrDefault(x => x.LobbyID == status.LobbyUser.AccountID);
-                if (entry != null) foreach (var kvp in entry.Parameters) script.AppendFormat("     {0}={1};\n", kvp.Key, kvp.Value);
-            }
-            script.AppendLine("  }");
-        }
 
 
         public BattleContext GetContext()
