+++ Shared/LobbyClient/TasClient.cs
@@ -12,6 +12,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
+using PlasmaShared;
 using ZkData;
 using ZkData.UnitSyncLib;
 using Timer = System.Timers.Timer;
