+++ Shared/PlasmaShared/ISpringieService/PlayerTeam.cs
@@ -3,10 +3,11 @@ namespace PlasmaShared
     public class PlayerTeam
     {
         public int LobbyID;
+        public int TeamID;
         public int AllyID;
         public string Name;
         public bool IsSpectator;
         public bool IsIngame;
-        public int TeamID;
+        public string ScriptPassword;
     }
 }
