+++ ZeroKLobby/MicroLobby/BattleChatControl.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
 using LobbyClient;
+using PlasmaShared;
 using ZkData.UnitSyncLib;
 using ZeroKLobby.Lines;
 using ZkData;
